More work on z_scene_proc.c 31/36 matching, 3 non-matching, 2 not attempted

+++ src/actors/Bg_Ikana_Ray/z_bg_ikana_ray.c
@@ -40,7 +40,7 @@ void BgIkanaRay_Init(ActorBgIkanaRay* this, GlobalContext* ctxt) {
     Collision_InitCylinderWithData(ctxt, collision, (Actor*)this, &bgIkanaRayCylinderInit);
     Collision_CylinderMoveToActor((Actor*)this, &this->collision);
 
-    this->animatedTextures = (AnimatedTexture*)Lib_PtrSegToVirt((u32)object_ikana_obj_001228);
+    this->animatedTextures = (AnimatedTexture*)Lib_PtrSegToVirt(object_ikana_obj_001228);
 
     if (Actor_GetSwitchFlag(ctxt, this->base.variable & 0x7F) != 0) {
         BgIkanaRay_SetActivated(this);
